Have Link Behavior Follow Updated Admin Settings
The settings interface was updated so the sizes are part of the "link to" option rather than simply having it offer "file". This made it so the previous code for getting the image wouldn't get the full image on "file" & instead would fall back to linking to the attachment page when full size was set.

This change will make it so it corresponds to the admin options.